+++ src/main/java/com/varunsdave/personafinance/budget/controller/IncomeController.java
@@ -18,7 +18,7 @@ public class IncomeController {
     private final String TYPE = "income";
 
     @PostMapping("/account/{accountId}")
-    public ResponseEntity<Transaction> createIncome(@PathVariable String accountId, @RequestBody int incomeAmt) {
+    public ResponseEntity<Transaction> createIncome(@PathVariable String accountId, @RequestBody double incomeAmt) {
        return ResponseEntity.ok().body( transactionService.create(incomeAmt, accountId, TYPE));
     }
